Декомпозирует редьюсер и добавит селекторы
